Create button functionality for human selection.
Added buttons for rock, paper, and scissors in
index.html. Each button had a variable created in
js file and then had eventlisteners added which
executed on a click and utlized callback function
to playRound(). Previous code for getting human
selection was deemed obsolote as the buttons are
now used for human selection.