Added functionality to update score and check who wins
updateDisplayedScore() function was created to
update score shown on html. The function was Added
to the playRound() function at the end to run each
time a player clicks a button.

checkWinner() function uses an if statement to
check when either the player or the computer reach
a score of 5 first. Whichever gets to 5 first is
displayed a message that indicates that they have
won.